Move CartWidget component to a new CartWidget directory
